+++ src/main/java/calculator/Delimiters.java
@@ -1,4 +1,4 @@
-package split;
+package calculator;
 
 import java.util.HashSet;
 import java.util.List;
