+++ src/main/java/calculator/StringCalculator.java
@@ -1,4 +1,4 @@
-package split;
+package calculator;
 
 import java.util.Objects;
 import java.util.regex.Matcher;
