+++ src/test/java/calculator/SetTest.java
@@ -1,4 +1,4 @@
-package split;
+package calculator;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
