+++ src/test/java/calculator/StringTest.java
@@ -1,4 +1,4 @@
-package split;
+package calculator;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
